refactor: 클립보드 복사 함수 유틸화
